+++ libs/sdk-core/src/lib/db/search/widget.ts
@@ -199,6 +199,7 @@ export namespace Widget {
   }
 
   export interface RaoWidgetConfiguration {
+    workflow: string;
     darkMode?: 'light' | 'dark';
   }