+++ libs/sdk-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support endpoint to trigger a sync job for a sync configuration
 - Add endpoints to get activity log for sync jobs
+- Support multiple workflows for Retrieval Agents
 
 # 1.31.1 (2026-01-20)